+++ tests/unit/test_validation.py
@@ -7,7 +7,7 @@ from typing import Callable
 
 import pytest
 
-from src.validation import verify_signature
+from webhook_router.validation import verify_signature
 from tests.unit.helpers import create_correct_signature, create_incorrect_signature
