+++ webhook_router/__init__.py
@@ -1,0 +1,2 @@
+#  Copyright 2024 Canonical Ltd.
+#  See LICENSE file for licensing details.
